Dead stores should be removed
Transifex client release does not work.
According to https://github.com/transifex/transifex-client/issues/113, the latest Transifex client does not work. Try to install from source in the meanwhile...

Convenience dev environment for NixOS

Travis, again

Remove the install step with Travis.
The install step is only needed when the language files must be updated. This is not required for PRs.

Use Python 3.5, upgrade pip and install the latest client.