Updating Transifex client used in Travis.
Simplify a bit the logic to install the Transifex client and push the generated language file.
Use API token set in Travis UI.
Install the latest transifex Python client and do not compile from source any more. Do this only when needed and not for every build. Should speed-up the build time a bit.