Tracking `devel-power-sockets` (#74)
* Power Sockets connect to Lamp Supplies

Had to also do a null check in InterSystem because of a null,null
Resistor that sits there while not connected.

* Fixup rendering and sync for power sockets

Sockets now have a color which dictates which cable color may connect to
them (as cables may be colored). Their "color care" is forced to
true--you cannot connect disparate colors to them. (Change the color
care on the wire instead.) The front face of the socket renders with a
lightened version of the color they will connect to.

Unrelated: the replicator maximum limit is now parameterized, but its
default remains the same.