Added Config Copy Tool. (#81)
* Added Config Copy Tool

Fixes #9.

Works mostly as intended, but there's a conflict in function with the
Wireless Button/Switch that causes it to not trigger for that block in
particular.

Currently supported:
- Lamp Supplies, Wireless Buttons/Switches, and Lamp Sockets copy their
  wireless/power channel names.
- Tachometers, Electrical Probes, Thermal Probes, Data Loggers, and
  Digital Displays copy their units, ranges, and periods, where it makes
  sense.

Other Elements can opt in by implementing the IConfigurable interface as
desired.

Sneak-right-click (or wrench-key-right-click) to copy, and normal
right-click to paste.

* Tool handles V sources and beeps

Voltage sources set the key "voltage"; right now, they only validly copy
between each other.

The tool now beeps when used. The beep gives a hint as to whether or not
the change worked: one kind of beep indicates that an upload was
successful, while another indicates that a download was successful; yet
a third indicates failure. The "failure" case is presently invoked on
NodeBase subclasses which don't implement
readConfigTool/writeConfigTool, or for those subclasses that do, when
the contained element does not implement IConfigurable. This is fairly
coarse, and does not necessarily mean that any changes were made--or
that the element actually had any data available to read, for example.

* Added inventory support to config tool

Plenty more items can now be configured--including lamp sockets, power
inductos, and power capacitors, which will sync to the number of items
copied out of your inventory, stopping if it runs out.

The invoker (an EntityPlayer) is now available to IConfigurable.

More targets still were added, and more are coming soon.

Here's a pony: https://derpicdn.net/img/view/2013/12/5/490099.jpeg

* Bypass limits in creative mode